Add demo mode (#174)
* Serve fake names from db

* Add env varible for demo mode

* Added fake name for demo purposes

* Moved setting of VITE_DEMO_MODE to compose file

As we only want to set DEMO MODE to true in the context of using the
whole project, it doesn't make sense using this from the Dockerfile. As
the student infos are already temporary placeholders in
frontend/server/index.js, I don't see it necessary to use this ENV
variable when building just this container.

* Disable react-in-jsx-scope rule & format

* Better approach to dropping the rule

---------

Co-authored-by: Joonas Tuominen <joonas.o.tuominen@outlook.com>